fix eslint, rename pincode file
